benchmark_data_loading/pyproject.toml
[tool.poetry]
name = "how-much-climate-change"
version = "0.1.0"
description = ""
authors = ["ali-ramadhan <ali.hh.ramadhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.12"}
tqdm = "^4.66.1"
xarray = "^2023.10.1"
netcdf4 = "^1.6.5"
pandas = "^2.1.3"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
joblib = "^1.3.2"
dask = "^2023.11.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
